Initial version
Added initial code to demonstrate the principle of
Inversion of Control (IoC).

In this version, our application uses two different
classes that implement a common interface, and the
application is responsible for instantiating the
concrete implementation. The problem is: we are
hardcoding the instantiation.  Ideally, we want to avoid
this.